Refactored function of game_on_going to check_game_state
