+++ GameAssistant/views/api/subclient.py
@@ -9,10 +9,11 @@ from asgiref.sync import async_to_sync
 from GameAssistant.models.clients import Client
 from GameAssistant.models.subclients import SubClient
 from GameAssistant.models.games import Game
-from GameAssistant.libs.utils import check_auth, game_ongoing, user_is_seated
+from GameAssistant.libs.utils import user_is_seated
+from GameAssistant.libs.utils_precheck import check_auth, check_game_state
 from GameAssistant.libs.utils_session import get_client_id_from_session, get_user_id_from_session
 from GameAssistant.libs.utils_websocket import ws_push
-from GameAssistant.libs.enums import SeatState, RefreshType
+from GameAssistant.libs.enums import GameState, SeatState, RefreshType
 import re
 
 @check_auth('guest')
@@ -68,7 +69,7 @@ def enter(request):
         return HttpResponseBadRequest('Unknown error while running subclient.enter! Details: {0}'.format(e))
 
 @ws_push('refeshing', RefreshType.seat.value)
-@game_ongoing('yes', 'subuser')
+@check_game_state(GameState.preparing.value, 'subuser')
 def sit(request):
     if request.method != 'POST':
         return HttpResponseBadRequest('Only POST are allowed!')
@@ -98,7 +99,7 @@ def sit(request):
         return HttpResponseBadRequest('Unknown error while running subclient.sit! Details: {0}'.format(e))
 
 @ws_push('refeshing', RefreshType.seat.value)
-@game_ongoing('yes', 'subuser')
+@check_game_state(GameState.preparing.value, 'subuser')
 def unsit(request):
     if request.method != 'POST':
         return HttpResponseBadRequest('Only POST are allowed!')
@@ -130,7 +131,7 @@ def unsit(request):
 
 
 @ws_push('refeshing', RefreshType.seat.value)
-@game_ongoing('yes', 'subuser')
+@check_game_state(GameState.preparing.value, 'subuser')
 def edit(request):
     if request.method != 'POST':
         return HttpResponseBadRequest('Only POST are allowed!')
