+++ GameAssistant/views/sites/going.py
@@ -1,17 +1,19 @@
+# -*- coding: utf-8 -*-
 from django.http import HttpResponse,HttpResponseRedirect,HttpResponseBadRequest
 from django.urls import reverse
 from django.contrib.sessions.models import Session
 from GameAssistant.models.clients import Client
 from GameAssistant.models.games import Game
-from GameAssistant.libs.utils import check_auth, game_ongoing
+from GameAssistant.libs.utils_precheck import check_game_state
+from GameAssistant.libs.enums import GameState
 from django.shortcuts import render
 
-@game_ongoing('yes', 'superuser')
+@check_game_state(GameState.preparing.value, 'superuser')
 def room(request):
     return render(request, "room_superuser.html")
 
 
-@game_ongoing('yes', 'subuser')
+@check_game_state(GameState.preparing.value, 'subuser')
 def room_guest(request):
     return render(request, "room_subuser.html")
